216: make findUpperAndLowerPoints() its own function
