+++ .eslintrc
@@ -32,6 +32,7 @@
     }],
     "no-param-reassign": "off",
     "no-lonely-if": "off",
+    "import/prefer-default-export": "off",
   },
   // these are globals made available through either the tagpro API or by
   // requiring them in src/index.js
